Made it so you can no longer destroy players by clicking on them
You can no longer destroy players when you click on them, now if you
click on a player it says "A Player was clicked on" To all players in
the server. Also i have added a new debug message, this message tells
everyone in the server if an object is destroyed by beaning clicked on
and also says the name of the object that was destroyed.