Added health hud for both players, and health variable
There is a nbew public variable called health that can be edited
directly in the code as well as in blueprints. The hud also now runs
fully off of the blue print system as well now so things can be added
and changed in the content editor side without having to touch the code.

Next i will be making it so the players can damage each other, add more
realistic hit detection including head shots and make it so that the
player can repawn.