Idle_walk blend space
The states for idle and walk in the state machine in the
PlayerAnimation/AnimGraph were replaced with a idle_walk state that
sends the character's velocity vector magnitude to the idle_walk blend
space. The blend space has the RobotMan_Anim_Skeleton_Idle animation at
X=0, and the RobotMan_Anim_Skeleton_Walk animation placed at X=375. The
private boolean variable running, and where it was used in the
blueprints, were removed.